refactor: move adapter boundary vocabularies into application/ports.py
