+++ docs/how-to-implement-adapter.md
@@ -14,9 +14,8 @@ Both live in `delta_engine.application.ports`. Both are `runtime_checkable` Prot
 ### CatalogStateReader
 
 ```python
-from delta_engine.application.ports import CatalogStateReader
+from delta_engine.application.ports import CatalogStateReader, CatalogState, ReadFailed, TableAbsent, TablePresent
 from delta_engine.application.failures import ReadFailure
-from delta_engine.application.results import CatalogState, ReadFailed, TableAbsent, TablePresent
 from delta_engine.domain.model import QualifiedName
 
 class MyReader:
@@ -40,9 +39,8 @@ class MyReader:
 ### PlanExecutor
 
 ```python
-from delta_engine.application.ports import PlanExecutor
+from delta_engine.application.ports import PlanExecutor, ExecutionSummary, ExecutionSucceeded, ExecutionFailed
 from delta_engine.application.failures import ExecutionFailure
-from delta_engine.application.results import ExecutionSummary, ExecutionSucceeded, ExecutionFailed
 from delta_engine.domain.model import QualifiedName
 from delta_engine.domain.plan.actions import ActionPlan
