+++ src/delta_engine/application/engine.py
@@ -34,14 +34,17 @@ import logging
 from delta_engine.application.dependency_resolution import resolve
 from delta_engine.application.errors import SyncFailedError
 from delta_engine.application.failures import Failure
-from delta_engine.application.ports import CatalogStateReader, PlanExecutor
+from delta_engine.application.ports import (
+    CatalogState,
+    CatalogStateReader,
+    ExecutionSummary,
+    PlanExecutor,
+    ReadFailed,
+    TablePresent,
+)
 from delta_engine.application.registry import Registry
 from delta_engine.application.results import (
-    CatalogState,
-    ExecutionSummary,
-    ReadFailed,
     SyncReport,
-    TablePresent,
     TableRunReport,
 )
 from delta_engine.application.validation import validate_plan
